Add a new product selector component
