+++ _inc/client/plans/single-product-backup.jsx
@@ -23,11 +23,6 @@ export function SingleProductBackup( { plan, products, upgradeLinks, isFetchingD
 
 	return (
 		<React.Fragment>
-			<h1 className="plans-section__header">{ __( 'Single Products' ) }</h1>
-			<h2 className="plans-section__subheader">
-				{ __( "Just looking for backups? We've got you covered." ) }
-			</h2>
-
 			<div className="plans-section__single-product">
 				{ isFetchingData ? (
 					<div className="plans-section__single-product-skeleton is-placeholder" />
